+++ cfdm/mixin/netcdf.py
@@ -4683,7 +4683,7 @@ class NetCDFAggregation(NetCDFMixin):
         """Remove the netCDF aggregated_data terms.
 
         The aggregated data terms define the names of the fragment
-        array variables, and are stored in a netCDF file in an
+        array variables, as would be stored in a netCDF file in an
         "aggregated_data" attribute.
 
         .. versionadded:: (cfdm) 1.12.0.0
@@ -4702,20 +4702,20 @@ class NetCDFAggregation(NetCDFMixin):
         **Examples**
 
         >>> f.nc_set_aggregated_data(
-        ...     {'shape': 'shape',
-        ...      'location': 'location',
-        ...      'address': 'address'}
+        ...     {'map': 'fragment_map',
+        ...      'uris': 'fragment_uris',
+        ...      'identifiers': 'fragment_identifiers'}
         ... )
         >>> f.nc_has_aggregated_data()
         True
         >>> f.nc_get_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_del_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_has_aggregated_data()
         False
         >>> f.nc_del_aggregated_data()
@@ -4723,7 +4723,7 @@ class NetCDFAggregation(NetCDFMixin):
         >>> f.nc_get_aggregated_data()
         {}
         >>> f.nc_set_aggregated_data(
-        ...     'shape: shape, location: location address: address'
+        ...     'map: fragment_map, uris: fragment_uris identifiers: fragment_idenfiers'
         ... )
 
         """
@@ -4756,20 +4756,20 @@ class NetCDFAggregation(NetCDFMixin):
         **Examples**
 
         >>> f.nc_set_aggregated_data(
-        ...     {'shape': 'shape',
-        ...      'location': 'location',
-        ...      'address': 'address'}
+        ...     {'map': 'fragment_map',
+        ...      'uris': 'fragment_uris',
+        ...      'identifiers': 'fragment_identifiers'}
         ... )
         >>> f.nc_has_aggregated_data()
         True
         >>> f.nc_get_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_del_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_has_aggregated_data()
         False
         >>> f.nc_del_aggregated_data()
@@ -4777,7 +4777,7 @@ class NetCDFAggregation(NetCDFMixin):
         >>> f.nc_get_aggregated_data()
         {}
         >>> f.nc_set_aggregated_data(
-        ...     'shape: shape, location: location address: address'
+        ...     'map: fragment_map, uris: fragment_uris identifiers: fragment_idenfiers'
         ... )
 
         """
@@ -4809,20 +4809,20 @@ class NetCDFAggregation(NetCDFMixin):
         **Examples**
 
         >>> f.nc_set_aggregated_data(
-        ...     {'shape': 'shape',
-        ...      'location': 'location',
-        ...      'address': 'address'}
+        ...     {'map': 'fragment_map',
+        ...      'uris': 'fragment_uris',
+        ...      'identifiers': 'fragment_identifiers'}
         ... )
         >>> f.nc_has_aggregated_data()
         True
         >>> f.nc_get_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_del_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_has_aggregated_data()
         False
         >>> f.nc_del_aggregated_data()
@@ -4830,7 +4830,7 @@ class NetCDFAggregation(NetCDFMixin):
         >>> f.nc_get_aggregated_data()
         {}
         >>> f.nc_set_aggregated_data(
-        ...     'shape: shape, location: location address: address'
+        ...     'map: fragment_map, uris: fragment_uris identifiers: fragment_idenfiers'
         ... )
 
         """
@@ -4871,20 +4871,20 @@ class NetCDFAggregation(NetCDFMixin):
         **Examples**
 
         >>> f.nc_set_aggregated_data(
-        ...     {'shape': 'shape',
-        ...      'location': 'location',
-        ...      'address': 'address'}
+        ...     {'map': 'fragment_map',
+        ...      'uris': 'fragment_uris',
+        ...      'identifiers': 'fragment_identifiers'}
         ... )
         >>> f.nc_has_aggregated_data()
         True
         >>> f.nc_get_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_del_aggregated_data()
-        {'shape': 'shape',
-         'location': 'location',
-         'address': 'address'}
+        {'map': 'fragment_map',
+         'uris': 'fragment_uris',
+         'identifiers': 'fragment_identifiers'}
         >>> f.nc_has_aggregated_data()
         False
         >>> f.nc_del_aggregated_data()
@@ -4892,7 +4892,7 @@ class NetCDFAggregation(NetCDFMixin):
         >>> f.nc_get_aggregated_data()
         {}
         >>> f.nc_set_aggregated_data(
-        ...     'shape: shape, location: location address: address'
+        ...     'map: fragment_map, uris: fragment_uris identifiers: fragment_idenfiers'
         ... )
 
         """
@@ -4916,9 +4916,9 @@ class NetCDFAggregation(NetCDFMixin):
         :Returns:
 
             `str` or `None`
-                The removed fragment type, either ``'location'`` for
-                fragment files, or ``'value'`` for fragment unique
-                values, or `None` if no fragment type was set.
+                The removed fragment type, either ``'uri'`` for
+                fragment datasets, or ``'unique_value'`` for fragment
+                unique values, or `None` if no fragment type was set.
 
         """
         return self._nc_del("aggregation_fragment_type", None)
@@ -4930,9 +4930,10 @@ class NetCDFAggregation(NetCDFMixin):
 
         :Returns:
 
-            `str`
-                The fragment type, either ``'location'`` for fragment
-                files, or ``'value'`` for fragment unique values.
+            `str` or `None`
+                The fragment type, either ``'uri'`` for fragment
+                datsets, or ``'unique_value'`` for fragment unique
+                values, or `None` for an unspecified fragment type.
 
         """
         return self._nc_get("aggregation_fragment_type", None)
@@ -4945,9 +4946,9 @@ class NetCDFAggregation(NetCDFMixin):
         :Parameters:
 
             value: `str` or `None`
-                The fragment type, either ``'location'`` for fragment
-                files, ``'value'`` for fragment unique values, or
-                `None` for an unspecified fragment type.
+                The fragment type, either ``'uri'`` for fragment
+                files, ``'unique_value'`` for fragment unique values,
+                or `None` for an unspecified fragment type.
 
         :Returns: